Add username submission handling in header component
This commit adds an event handler that sets the username when the user finishes typing in the header component input box. The event gets triggered either by user focus out (blur event) or a press of the 'Enter' key. The username state management is now managed by a state subscriber in the header component rather than in the ngOnInit method. Additionally, this introduces a new method in the backend service to persist the username change.